refactor tabs designs
