+++ src/plugins/data_source_management/public/components/data_source_home_panel/data_source_page_header.tsx
@@ -30,7 +30,7 @@ export const DataSourceHeader: React.FC<DataSourceHeaderProps> = ({ history }) =
           <p>
             <FormattedMessage
               id="dataSourcesManagement.dataSourcesTable.description"
-              defaultMessage="Create and manage data source connections to help you retrieve data from multiple OpenSearch compatible sources."
+              defaultMessage="Create and manage data source connections."
             />
           </p>
         </EuiText>
