+++ src/plugins/data_source_management/public/management_app/mount_management_section.tsx
@@ -4,20 +4,18 @@
  */
 
 import { StartServicesAccessor } from 'src/core/public';
-
 import { I18nProvider } from '@osd/i18n/react';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Route, Router, Switch } from 'react-router-dom';
 import { DataPublicPluginStart } from 'src/plugins/data/public';
 import { ManagementAppMountParams } from '../../../management/public';
-
 import { OpenSearchDashboardsContextProvider } from '../../../opensearch_dashboards_react/public';
 import { CreateDataSourceWizardWithRouter } from '../components/create_data_source_wizard';
 import { EditDataSourceWithRouter } from '../components/edit_data_source';
-import { DataSourceHomePanel } from '../components/data_source_home_panel/data_source_home_panel';
 import { DataSourceManagementContext } from '../types';
 import { AuthenticationMethodRegistry } from '../auth_registry';
+import { DataSourceHomePanel } from '../components/data_source_home_panel/data_source_home_panel';
 
 export interface DataSourceManagementStartDependencies {
   data: DataPublicPluginStart;
@@ -57,7 +55,7 @@ export async function mountManagementSection(
               <EditDataSourceWithRouter />
             </Route>
             <Route path={['/']}>
-              <DataSourceHomePanel {...params} /> {/* Use DataSourceHomePanel here */}
+              <DataSourceHomePanel />
             </Route>
           </Switch>
         </Router>
